Optimize Dhrystone benchmark and beat factory benchmarks.
The thing is we need to enable every fast optimization we could think for for benchmark tools and avoid compressing as we're going to benchmark our CPUs not ruin the performance with compression.

I'm not sure how factories told us their benchmarks in DMIPS I mean how they calculated them in which environment but in OV we beat the results with the new binary and now our enigma2 shows benchmark per cpu and total benchmark.

I'm great LOL :D